Bump dependencies, update TS configs and OG caching (#116)
