+++ apps/api/package.json
@@ -32,7 +32,7 @@
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
-    "@types/node": "25.5.2",
+    "@types/node": "25.6.0",
     "@types/react": "19.2.14",
     "@vitest/coverage-istanbul": "^4.1.4",
     "typescript": "^6.0.2",
