+++ apps/email/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "@repo/typescript-config/nextjs.json",
-  "include": ["**/*.ts", "**/*.tsx"],
+  "files": [],
   "exclude": ["node_modules"]
 }
