+++ apps/www/app/og/[...slug]/route.tsx
@@ -1,13 +1,8 @@
 import { routing } from "@repo/internationalization/src/routing";
-import { Effect } from "effect";
 import type { NextRequest } from "next/server";
 import { hasLocale, type Locale } from "next-intl";
-import { generateOGImage } from "@/app/[locale]/og/[...slug]/og";
-import { getMetadataFromSlug } from "@/lib/utils/system";
-
-export function generateStaticParams() {
-  return [];
-}
+import { generateOGImage } from "@/lib/og";
+import { getCachedMetadataFromSlug } from "@/lib/utils/system";
 
 export async function GET(
   _req: NextRequest,
@@ -25,11 +20,12 @@ export async function GET(
   const contentSlug =
     cleanSlug.at(-1) === "image.png" ? cleanSlug.slice(0, -1) : cleanSlug;
 
-  const { title, description } = await Effect.runPromise(
-    getMetadataFromSlug(locale, contentSlug)
+  const { title, description } = await getCachedMetadataFromSlug(
+    locale,
+    contentSlug
   );
 
-  return generateOGImage({
+  return await generateOGImage({
     title,
     description,
   });
