+++ apps/www/package.json
@@ -19,7 +19,7 @@
   },
   "dependencies": {
     "@ai-sdk/devtools": "^0.0.15",
-    "@ai-sdk/react": "^3.0.156",
+    "@ai-sdk/react": "^3.0.160",
     "@convex-dev/better-auth": "^0.11.4",
     "@hugeicons/core-free-icons": "^4.1.1",
     "@hugeicons/react": "^1.1.6",
@@ -28,7 +28,7 @@
     "@mantine/hooks": "^9.0.1",
     "@mdx-js/loader": "^3.1.1",
     "@mdx-js/react": "^3.1.1",
-    "@paper-design/shaders-react": "^0.0.72",
+    "@paper-design/shaders-react": "^0.0.73",
     "@repo/ai": "workspace:*",
     "@repo/analytics": "workspace:*",
     "@repo/backend": "workspace:*",
@@ -43,19 +43,18 @@
     "@t3-oss/env-nextjs": "^0.13.11",
     "@tabler/icons-react": "^3.41.1",
     "@tailwindcss/postcss": "^4.2.2",
-    "@takumi-rs/core": "^0.73.1",
-    "@takumi-rs/image-response": "^0.73.1",
-    "@tanstack/react-form": "^1.28.6",
-    "@tanstack/react-query": "^5.97.0",
-    "@tanstack/react-query-devtools": "^5.97.0",
+    "@takumi-rs/image-response": "^1.0.5",
+    "@tanstack/react-form": "^1.29.0",
+    "@tanstack/react-query": "^5.99.0",
+    "@tanstack/react-query-devtools": "^5.99.0",
     "@vercel/functions": "^3.4.3",
-    "ai": "^6.0.154",
+    "ai": "^6.0.158",
     "any-ascii": "^0.3.3",
     "babel-plugin-react-compiler": "^1.0.0",
     "better-auth": "1.5.3",
     "class-variance-authority": "^0.7.1",
-    "convex": "^1.34.1",
-    "country-flag-icons": "^1.6.15",
+    "convex": "^1.35.1",
+    "country-flag-icons": "^1.6.16",
     "date-fns": "^4.1.0",
     "effect": "^3.21.0",
     "feed": "^5.2.0",
@@ -66,7 +65,7 @@
     "motion": "^12.38.0",
     "nanoid": "^5.1.7",
     "next": "16.2.3",
-    "next-intl": "^4.9.0",
+    "next-intl": "^4.9.1",
     "next-themes": "^0.4.6",
     "nuqs": "^2.8.9",
     "pino": "^10.3.1",
@@ -76,14 +75,14 @@
     "sonner": "^2.0.7",
     "use-context-selector": "^2.0.0",
     "use-sound": "^5.0.0",
-    "virtua": "^0.49.0",
+    "virtua": "^0.49.1",
     "zod": "^4.3.6",
     "zustand": "^5.0.12"
   },
   "devDependencies": {
     "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@types/node": "25.5.2",
+    "@types/node": "25.6.0",
     "@types/react": "19.2.14",
     "@types/react-dom": "19.2.3",
     "@vitest/coverage-istanbul": "^4.1.4",
