+++ apps/www/vitest.config.ts
@@ -1,1 +1,13 @@
-export { default } from "@repo/testing";
+import config from "@repo/testing";
+import { mergeConfig } from "vitest/config";
+
+export default mergeConfig(config, {
+  test: {
+    coverage: {
+      thresholds: {
+        100: true,
+        perFile: true,
+      },
+    },
+  },
+});
