+++ packages/analytics/package.json
@@ -9,7 +9,7 @@
   "dependencies": {
     "@t3-oss/env-nextjs": "^0.13.11",
     "@vercel/analytics": "^2.0.1",
-    "posthog-js": "^1.366.1",
+    "posthog-js": "^1.367.0",
     "posthog-node": "^5.29.2",
     "react": "^19.2.5",
     "server-only": "^0.0.1",
@@ -17,7 +17,7 @@
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
-    "@types/node": "25.5.2",
+    "@types/node": "25.6.0",
     "@types/react": "19.2.14",
     "@types/react-dom": "19.2.3"
   }
