+++ packages/backend/package.json
@@ -54,10 +54,10 @@
     "@repo/ai": "workspace:*",
     "@repo/utilities": "workspace:*",
     "@t3-oss/env-nextjs": "^0.13.11",
-    "ai": "^6.0.154",
+    "ai": "^6.0.158",
     "better-auth": "1.5.3",
     "buffer": "^6.0.3",
-    "convex": "^1.34.1",
+    "convex": "^1.35.1",
     "convex-helpers": "^0.1.114",
     "hono": "^4.12.12",
     "jsonrepair": "^3.13.3",
