+++ packages/contents/package.json
@@ -19,7 +19,7 @@
     "effect": "^3.21.0",
     "ky": "^2.0.0",
     "motion": "^12.38.0",
-    "next-intl": "^4.9.0",
+    "next-intl": "^4.9.1",
     "react": "19.2.5",
     "react-aria-components": "^1.16.0",
     "recharts": "^2.15.4",
@@ -27,7 +27,7 @@
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
-    "@types/node": "25.5.2",
+    "@types/node": "25.6.0",
     "@types/react": "^19.2.14",
     "@vitest/coverage-istanbul": "^4.1.4",
     "vitest": "^4.1.4"
