+++ packages/design-system/package.json
@@ -30,7 +30,7 @@
     "@tabler/icons-react": "^3.41.1",
     "@tailwindcss/postcss": "^4.2.2",
     "@types/mdx": "^2.0.13",
-    "ai": "^6.0.154",
+    "ai": "^6.0.158",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -50,7 +50,7 @@
     "motion": "^12.38.0",
     "nanoid": "^5.1.7",
     "next": "^16.2.3",
-    "next-intl": "^4.9.0",
+    "next-intl": "^4.9.1",
     "next-themes": "^0.4.6",
     "postcss": "^8.5.9",
     "radix-ui": "^1.4.3",
@@ -64,7 +64,7 @@
     "react-moveable": "^0.56.0",
     "react-nice-avatar": "^1.5.0",
     "react-phone-number-input": "^3.4.16",
-    "react-resizable-panels": "^4.9.0",
+    "react-resizable-panels": "^4.10.0",
     "react-syntax-highlighter": "^16.1.1",
     "react-textarea-autosize": "^8.5.9",
     "recharts": "^2.15.4",
@@ -83,13 +83,13 @@
     "use-context-selector": "^2.0.0",
     "use-stick-to-bottom": "^1.1.3",
     "vaul": "^1.1.2",
-    "virtua": "^0.49.0"
+    "virtua": "^0.49.1"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
     "@tailwindcss/typography": "^0.5.19",
     "@types/lodash.throttle": "^4.1.9",
-    "@types/node": "^25.5.2",
+    "@types/node": "^25.6.0",
     "@types/react": "^19.2.14",
     "@types/react-katex": "^3.0.4",
     "@types/react-syntax-highlighter": "^15.5.13",
