+++ packages/seo/package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "lodash.merge": "^4.6.2",
-    "next-intl": "^4.9.0",
+    "next-intl": "^4.9.1",
     "react": "^19.2.5",
     "schema-dts": "^2.0.0",
     "zod": "^4.3.6"
